requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.28
websockets==12.0
python-dotenv==1.0.1
chromadb==0.4.24
Pillow==10.2.0
numpy==1.26.4
beautifulsoup4==4.12.3
httpx==0.27.0
loguru==0.7.2
openai==1.16.1
tenacity==8.2.3
